new header file

+++ 0x00-math_complex/holberton.h
@@ -11,8 +11,8 @@
  */
 typedef struct Complex
 {
-	double re;
-	double im;
+	int re;
+	int im;
 } complex;
 
 /* functions prototypes */
